requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7

python-dotenv==1.0.1

transformers==4.44.2
torch>=2.2.0
sentence-transformers==3.0.1
numpy==1.24.3
pandas==2.2.2
joblib==1.4.2

pdfminer.six==20231228
PyPDF2==3.0.1

ujson==5.10.0
